separates redis oob contexts in separate source files to facilitate cython processing
